// src/types.ts
export interface Petani {
  id: string;
  nama: string;
  nik: string;
  noHp: string;
  blokLahan: string;
  luasHa: number;
  jmlPokok: number;
  noRekening: string;
  bank: string;
  statusAktif: boolean;
  tanggalGabung: string;
  fotoUrl?: string;
}
// ...
export interface PanenRecord {
  id: string;
  noSpb: string; // Surat Pengantar Buah
  tanggal: string;
  petaniId: string;
  petaniNama: string;
  blokLahan: string;
  
  // Tonase
  timbanganRamKg: number;    // Berat timbangan kebun / ram luar
  timbanganPksKg: number;    // Berat netto timbangan pabrik
  selisihKg: number;         // timbanganRamKg - timbanganPksKg (penyusutan)
  persentaseSelisih: number; // (selisihKg / timbanganRamKg) * 100
  
  // Harga & Bruto
  hargaTbsPerKg: number;     // Rp/Kg
  totalBruto: number;        // timbanganPksKg * hargaTbsPerKg
  
  // Potongan / Pedaran
  potonganPedaranKg: number;     // Potongan sortasi / sampah / tangkai panjang
  potonganPedaranRupiah: number; // potonganPedaranKg * hargaTbsPerKg
  potonganIuranKasRupiah: number;// Misal Rp 25/kg dari timbangan PKS
  upahPemanenRupiah: number;     // Misal Rp 180/kg
  kasbonPupukRupiah: number;     // Potongan angsuran kasbon/pupuk
  totalPotongan: number;         // Total seluruh potongan
  
  // Netto Petani (Timbangan Kebun × Harga TBS)
  totalNetto: number;            // (timbanganRamKg || timbanganPksKg) * hargaTbsPerKg
  
  // Info Logistik & PKS
  namaPks: string;               // Contoh: PKS PT. Sawit Sejahtera, PKS Agro Mandiri
  platTruk: string;              // BM 8412 TA
  namaSopir: string;
  namaPemanen: string;
  
  // Status
  statusPembayaran: 'Lunas' | 'Siap Bayar' | 'Draft';
  tanggalBayar?: string;
  metodeBayar?: 'Transfer Bank' | 'Tunai / Kas';
  catatan?: string;
}